[PI-1][S1][Project Setup] - updated the NSN project structure.
