+++ services/nsn/modules/java/nsn-routing-service/pom.xml
@@ -4,13 +4,11 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>gov.gsa.fss19</groupId>
-  <artifactId>nsnmaintenanceservice</artifactId>
+  <groupId>gov.gsa.fas</groupId>
+  <artifactId>nsn-routing-service</artifactId>
   <version>1.0-SNAPSHOT</version>
 
-  <name>nsnmaintenanceservice</name>
-  <!-- FIXME change it to the project's website -->
-  <url>http://www.example.com</url>
+  <name>nsn-routing-service</name>
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
