+++ services/nsn/modules/java/nsn-routing-service/src/main/java/gov/gsa/fas/nsn/handler/NSNDataPutHandler.java
@@ -1,35 +1,29 @@
-package gov.gsa.fss19.nsn.handler;
+package gov.gsa.fas.nsn.handler;
 
-import java.io.BufferedReader;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.io.OutputStream;
-import java.nio.charset.Charset;
 import java.util.Map;
 
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
 import com.amazonaws.services.dynamodbv2.document.DynamoDB;
 import com.amazonaws.services.dynamodbv2.document.Item;
-import com.amazonaws.services.dynamodbv2.document.KeyAttribute;
 import com.amazonaws.services.dynamodbv2.document.PrimaryKey;
 import com.amazonaws.services.dynamodbv2.document.Table;
+import com.amazonaws.services.dynamodbv2.document.spec.DeleteItemSpec;
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.LambdaLogger;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
 import com.amazonaws.util.StringUtils;
 import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
 
-import gov.gsa.fss19.model.GatewayResponse;
-import gov.gsa.fss19.model.NSNData;
+import gov.gsa.fas.nsn.model.GatewayResponse;
+import gov.gsa.fas.nsn.model.NSNData;
 
-public class NSNDataPostHandler implements RequestHandler<Map<String, Object>, GatewayResponse> {
+public class NSNDataPutHandler implements RequestHandler<Map<String, Object>, GatewayResponse>  {
 	
 	static AmazonDynamoDB client = AmazonDynamoDBClientBuilder.standard().build();
     static DynamoDB dynamoDB = new DynamoDB(client);
     static String tableName = "NSN_DATA";
-    
+
     @Override
 	public GatewayResponse handleRequest(Map<String, Object> inputDataMap, Context context) {
     	LambdaLogger logger = context.getLogger();
@@ -40,14 +34,19 @@ public class NSNDataPostHandler implements RequestHandler<Map<String, Object>, G
     	NSNData inputData = gson.fromJson(inputDataStr, NSNData.class);
     	
     	Table table = dynamoDB.getTable(tableName);
-    	
     	Item existingItem = table.getItem("NSN_ID", inputData.getRoutingId());
     	
-    	if(existingItem != null) {
-    		return new GatewayResponse("NSN data exists with ID - "+inputData.getRoutingId(), 400, null, false);
-    	}
+    	if(existingItem == null) {
+    		return new GatewayResponse("No NSN data exists for routing id - "+inputData.getRoutingId(), 404, null, false);
+		}else {
+			DeleteItemSpec deleteItemSpec = new DeleteItemSpec()
+					.withPrimaryKey(new PrimaryKey("NSN_ID", inputData.getRoutingId()));
+			table.deleteItem(deleteItemSpec);
+		}
     	
     	Item item = new Item().withPrimaryKey("NSN_ID", inputData.getRoutingId());
+		
+		
 		if(!StringUtils.isNullOrEmpty(inputData.getIsCivMgr())) {
 			item = item.withString("CIV_MGR", inputData.getIsCivMgr().equalsIgnoreCase("Y") ? "true" : "false");
 		}
@@ -68,6 +67,6 @@ public class NSNDataPostHandler implements RequestHandler<Map<String, Object>, G
 		}
 		table.putItem(item);
 		
-		return new GatewayResponse("NSN Routing record created for ID - "+inputData.getRoutingId(), 200, null, false);
-	}	
+		return new GatewayResponse("NSN data updated for routing id - "+inputData.getRoutingId(), 200, null, false);
+	}
 }
