--- services/nsn/src/main/java/gov/gsa/fss19/App.java
@@ -1,13 +1,0 @@
-package gov.gsa.fss19;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
