--- services/nsn/src/test/java/gov/gsa/fss19/AppTest.java
@@ -1,20 +1,0 @@
-package gov.gsa.fss19;
-
-import static org.junit.Assert.assertTrue;
-
-import org.junit.Test;
-
-/**
- * Unit test for simple App.
- */
-public class AppTest 
-{
-    /**
-     * Rigorous Test :-)
-     */
-    @Test
-    public void shouldAnswerWithTrue()
-    {
-        assertTrue( true );
-    }
-}
